Removed warning for unsigned int compare.

+++ tests/create_low_rank_array.h
@@ -12,12 +12,12 @@ Matrix<T> low_rank_matrix(int rows, int cols, int rank) {
     auto full_rank = std::min(rows, cols);
     assert(rank <= std::min(rows, cols));
 
-    Eigen::JacobiSVD
-        <Matrix<T>> svd(Matrix<T>::Random(rows, cols), Eigen::ComputeThinU
-                        | Eigen::ComputeThinV);
+    Eigen::JacobiSVD<Matrix<T>> svd(Matrix<T>::Random(rows, cols),
+                                    Eigen::ComputeThinU | Eigen::ComputeThinV);
     auto svals = svd.singularValues();
 
-    for (auto i = 0ul; i < (full_rank - rank) && i < svals.size(); ++i) {
+    for (auto i = 0;
+         i < (full_rank - rank) && i < static_cast<int>(svals.size()); ++i) {
         svals[(svals.size() - 1) - i] = 0;
     }
